added view file for dependency inversion
